requirements.txt
altair==5.4.1
attrs==24.2.0
backports.zoneinfo==0.2.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.55.0
gitdb==4.0.11
GitPython==3.1.43
h5py==3.11.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
narwhals==1.14.0
numpy==1.24.4
openpyxl==3.1.5
packaging==23.2
pandas==2.0.3
photovoltaic==0.1.9
pillow==10.4.0
pip==24.2
pkgutil_resolve_name==1.3.10
protobuf==4.25.5
pvlib==0.11.0
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
scipy==1.10.1
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
streamlit==1.39.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
watchdog==4.0.2
wheel==0.44.0
xlrd==2.0.1
zipp==3.20.2